split out getTranslateFunction behavior
