+++ src/index.js
@@ -7,3 +7,4 @@ export default from './component'
 export {i18nState} from './reducer'
 export {setLanguage, setTranslations} from './actions'
 export localize from './hoc'
+export getTranslateFunction from './getTranslateFunction';
